requirements.txt
djangorestframework-jwt
asgiref==3.2.3
bson==0.5.8
dataclasses==0.6
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
djongo==1.3.0
dnspython==1.16.0
Markdown==3.2
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
text-unidecode==1.3
